1. Dinh dang lai so dien thoai  2. Validate email, so dien thoai 3. Them dialog cho delete button   4. Them dialog-delete-confirm component

+++ demo/src/app/page/employees/employees.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { EmployeeService } from '../../services/employee.service';
 import { Employee } from '../../models/employee.model';
+import { ConfirmDialogService } from '../../services/confirm-dialog.service';
 
 @Component({
   selector: 'app-employees',
@@ -9,14 +10,19 @@ import { Employee } from '../../models/employee.model';
 })
 export class EmployeesComponent implements OnInit {
   employees: Employee[] = [];
-  constructor(private employeeService: EmployeeService) { }
+  constructor(
+    private employeeService: EmployeeService,
+    private confirmDialogService: ConfirmDialogService
+  ) { }
 
   ngOnInit(): void {
     this.employees = this.employeeService.onGet();
   }
   onDelete(id: number){
-    window.alert("Are you sure ?");
-    this.employeeService.onDelete(id);
+    let confirm = window.confirm('Nhấp OK nếu bạn chắc chắn muốn xóa?');
+    if (confirm){
+      this.employeeService.onDelete(id);
+    }
   }
 
   formatDate(date: Date){
@@ -24,5 +30,12 @@ export class EmployeesComponent implements OnInit {
     let mm = String(date.getMonth() + 1).padStart(2, '0');
     let yyyy = String(date.getFullYear()).padStart(4, '0');
     return dd + '-' + mm + '-' + yyyy;
+  }
+  showDialog(){
+    this.confirmDialogService.confirmThis('Are you sure to delete ?', function(){
+      alert('Yes');
+    }, function(){
+      alert('No');
+    })
   }
 }
